Added Some Enemy Tanks, Debugging Route Finding Code
* Threw together some quick enemy tank prefabs and put them on the map for testing purposes. An enemy faction to test route finding against and to see how different movement types work. Plus it makes the map feel a bit more complete. (lol it is nowhere near complete)

* New asset so enemy spawn prefabs have a red circle instead of the player blue one

* Fixed the movement type of the enemy tanks so they are a ground vehicle instead of general

* Fixed a casting error with inhibition scores. They should be ulongs to keep things consistent but they were uints in some places. Likely never will be a problem but still this has been corrected.

* Inhibition scores for tile groups is now being calculated. However they don't seem to be having an impact on route finding at the moment, need to look into that.

* i'm an idiot with how factions are being handled atm, need to review how I want to do them

* Fixed a bug with tile movement costs. The passibility of the movement cost was never being considered, now it'll check it in the GameMapTileData when it checks the movement cost. If the movement cost is flagged as not passible it'll return the same NaN as if no movement data was found.

* Quick refactor: Changed the route overlap check to be done in reverse, from the end. It should find tiles that overlap (like the route immediately before the end) and remove it from the directions that need to be checked so it should run slightly faster.

* Removed old depth first route finding code since it is completely obsolete with the current breadth route finding code.

* Added Linq ForEach bit, didn't use it yet but it's handy to have

* Working on adding checks for where routes can't end, such as on top of another unit. Currently not working, need to rethink how I want to handle it. I should also check how I'm handling impassible tiles too, there's probably a decent optimization to be had here with reducing repeated work.

+++ Size Emblem Project/Assets/Scripts/Interfaces/GameMap/IGameMapTileGroup.cs
@@ -15,6 +15,6 @@ namespace SizeEmblem.Scripts.Interfaces.GameMap
         void SetLayerTile(int layer, IGameMapTile tile);
 
         float GetMovementCostForType(MovementType movementType);
-        uint GetInhibitionScoreForUnit(IGameUnit unit);
+        ulong GetInhibitionScoreForUnit(IGameUnit unit);
     }
 }
